requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

# Supabase
supabase>=2.10.0
gotrue>=2.12.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.2.4

# Redis & Celery
redis>=5.0.0
celery>=5.3.0

# Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# AI & Voice
openai>=1.12.0
deepgram-sdk>=3.1.0
anthropic>=0.18.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
structlog>=24.1.0

# Document Processing
pypdf2>=3.0.0
python-docx>=1.1.0
tiktoken>=0.5.0

# WebSocket
websockets>=13.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
